+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @snowballtools/auth
 
+## 0.2.0
+
+### Minor Changes
+
+- Upgrade lit to v6
+
+### Patch Changes
+
+- Updated dependencies
+  - @snowballtools/types@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes