+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @snowballtools/types
 
+## 0.2.0
+
+### Minor Changes
+
+- Upgrade lit to v6
+
 ## 0.1.1
 
 ### Patch Changes